* mu4e-overview.el (mu4e-overview-mode-map): Set parent map
Use `mu4e-main-mode-map' as the parent keymap, making it
unnecessary to ever go to the " *mu4e-main*" buffer just to
invoke certain commands.

Bind "q" to `bury-buffer' because the quit function from
`mu4e-main-mode-map' is not suitable here.

Because one might not always remember the key bindings bind "?"
to `mu4e~main-menu', making it only necessary to remember "?".